Fix bug with grow/shrink and wrong layout in the `ZeroAwareAllocator`
This fixes a bug where we could accidentally pass the wrong `Layout` to the
underlying allocator when growing or shrinking an allocation. When we have an
already-zeroed block of size N, we can satisfy allocations of size `M` where `M
< N` if we are willing to accept some fragmentation slop. We currently do this
but only when the slop is relatively small (1/8 the size of the allocation or
less). If we handed out such an allocation, and then the user wants to grow or
shrink it, then we would pass the user's layout to the underlying allocator,
rather than the original layout. Passing the wrong layout to an allocator is a
violation of the `Allocator` interface's safety contract, and could therefore
lead to UB. In fact, this was leading to assertion failures when paired with
`dlmalloc`.

This commit also sets up and enables fuzzing via libfuzzer with our existing
mutator and oracle; previously we were only using the mutator and oracle for
property-testing style tests. I forgot to do this originally and just stopped
after I had the property-tests working with the mutator and oracle.