Tweak a failing alignment assertion that did not always hold true
It is not necessarily the case that the original layout's alignment is greater
than or equal to the user's layout's alignment. We may have reused an allocation
that happened to be aligned greater than its original layout's requested
alignment such that it satisfied a new user layout's alignment. Therefore, we
instead assert that the pointer is aligned as expected.